Organize utils types into separate file
