+++ pkg/utils/types.go
@@ -1,0 +1,9 @@
+package utils
+
+func Bool(b bool) *bool {
+	return &b
+}
+
+func UInt64(u uint64) *uint64 {
+	return &u
+}
