+++ pkg/utils/utils.go
@@ -50,11 +50,3 @@ func Includes(slice []string, value string) bool {
 	}
 	return false
 }
-
-func Bool(b bool) *bool {
-	return &b
-}
-
-func UInt64(u uint64) *uint64 {
-	return &u
-}
